+++ R/RcppExports.R
@@ -17,6 +17,10 @@ annoy_manhattan_nns <- function(index_name, mat, n, search_k, grain_size = 1L, v
     .Call(`_uwot_annoy_manhattan_nns`, index_name, mat, n, search_k, grain_size, verbose)
 }
 
+annoy_hamming_nns <- function(index_name, mat, n, search_k, grain_size = 1L, verbose = FALSE) {
+    .Call(`_uwot_annoy_hamming_nns`, index_name, mat, n, search_k, grain_size, verbose)
+}
+
 optimize_layout_umap <- function(head_embedding, tail_embedding, positive_head, positive_tail, n_epochs, n_vertices, epochs_per_sample, a, b, gamma, initial_alpha, negative_sample_rate, seed, approx_pow, parallelize = TRUE, grain_size = 1L, move_other = TRUE, verbose = FALSE) {
     .Call(`_uwot_optimize_layout_umap`, head_embedding, tail_embedding, positive_head, positive_tail, n_epochs, n_vertices, epochs_per_sample, a, b, gamma, initial_alpha, negative_sample_rate, seed, approx_pow, parallelize, grain_size, move_other, verbose)
 }